requirements.txt
numpy==2.2.1
pillow==11.0.0
pycryptodome==3.21.0
